#229 refactor: check file exists.

+++ entrypoint.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
 # /appをホストにマウントする場合、jarファイルが消えてしまうのでコピーしておく (クライアント未作成時の起動などで利用される想定)
-cp /dbflute-intro.jar /app/dbflute-intro.jar
+if [ ! -e /app/dbflute-intro.jar ]; then
+  cp /dbflute-intro.jar /app/dbflute-intro.jar
+fi
 
 echo "launch dbflute-intro.jar..."
 java -jar -Dintro.host=0.0.0.0 dbflute-intro.jar
